test: Get Rid of OpenID Connect Bypass in API Tests
Previously, OpenID Connect was bypassed for API tests. This is a
possible security issue. The bypass is now removed.

Instead, the API test spawns a simple HTTP server which emulates the
behaviour of an OpenID Connect server. It provides a Discovery
Endpoint and an Inspection Endpoint.

Now, API test can run with OpenID Connect enabled.

Signed-off-by: Philipp Le <philipp@philipple.de>